Changed internal rep of signed and unsigned integer from long long and unsigned long long to int64_t and uint64_t

+++ src/jsoncons/json_parser.hpp
@@ -1857,7 +1857,7 @@ private:
         {
             try
             {
-                long long d = string_to_integer(is_negative_, number_buffer_.c_str(), number_buffer_.length());
+                int64_t d = string_to_integer(is_negative_, number_buffer_.c_str(), number_buffer_.length());
                 handler_->value(d, *this);
             }
             catch (const std::exception&)
@@ -1878,7 +1878,7 @@ private:
         {
             try
             {
-                unsigned long long d = string_to_unsigned_integer(number_buffer_.c_str(), number_buffer_.length());
+                uint64_t d = string_to_unsigned_integer(number_buffer_.c_str(), number_buffer_.length());
                 handler_->value(d, *this);
             }
             catch (const std::exception&)
@@ -2123,14 +2123,14 @@ private:
     }
 
     template<typename CharType>
-    unsigned long long string_to_unsigned_integer(const CharType *s, size_t length) throw(std::overflow_error)
+    uint64_t string_to_unsigned_integer(const CharType *s, size_t length) throw(std::overflow_error)
     {
-        static const unsigned long long max_value = std::numeric_limits<unsigned long long>::max JSONCONS_NO_MACRO_EXP();
-        static const unsigned long long max_value_div_10 = max_value / 10;
-        unsigned long long n = 0;
+        static const uint64_t max_value = std::numeric_limits<uint64_t>::max JSONCONS_NO_MACRO_EXP();
+        static const uint64_t max_value_div_10 = max_value / 10;
+        uint64_t n = 0;
         for (size_t i = 0; i < length; ++i)
         {
-            unsigned long long x = s[i] - '0';
+            uint64_t x = s[i] - '0';
             if (n > max_value_div_10)
             {
                 throw std::overflow_error("Unsigned overflow");
